+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - SliderAudioGroup component
 - TextMeshProAudioGroup component
 - ToggleAudioGroup component
+- AudioSourceAudioGroup component
 
 ### Changed
 - Remove DisallowMultipleComponent attribute from AudioSourceDictionary